[Refactor/#93] UserRank 도메인 모델 클래스 rank 패키지로 이
