+++ core/data/src/main/java/com/ilsangtech/ilsang/core/data/rank/mapper/RankMapper.kt
@@ -1,7 +1,7 @@
 package com.ilsangtech.ilsang.core.data.rank.mapper
 
 import com.ilsangtech.ilsang.core.data.title.mapper.toTitle
-import com.ilsangtech.ilsang.core.model.UserRank
+import com.ilsangtech.ilsang.core.model.rank.UserRank
 import com.ilsangtech.ilsang.core.network.model.rank.UserRankNetworkModel
 
 internal fun UserRankNetworkModel.toUserRank(): UserRank {
