+++ core/data/src/main/java/com/ilsangtech/ilsang/core/data/rank/repository/RankRepositoryImpl.kt
@@ -5,8 +5,8 @@ import com.ilsangtech.ilsang.core.data.rank.mapper.toUserRank
 import com.ilsangtech.ilsang.core.data.rank.toUserXpTypeRank
 import com.ilsangtech.ilsang.core.domain.RankRepository
 import com.ilsangtech.ilsang.core.model.RewardType
-import com.ilsangtech.ilsang.core.model.UserRank
 import com.ilsangtech.ilsang.core.model.UserXpTypeRank
+import com.ilsangtech.ilsang.core.model.rank.UserRank
 import com.ilsangtech.ilsang.core.network.model.rank.UserRankNetworkModel
 import com.ilsangtech.ilsang.core.network.model.rank.XpTypeRankNetworkModel
 import kotlinx.coroutines.flow.Flow
