+++ feature/home/src/main/java/com/ilsangtech/ilsang/feature/home/model/HomeTabUiState.kt
@@ -1,10 +1,10 @@
 package com.ilsangtech.ilsang.feature.home.model
 
-import com.ilsangtech.ilsang.core.model.UserRank
 import com.ilsangtech.ilsang.core.model.banner.Banner
 import com.ilsangtech.ilsang.core.model.quest.LargeRewardQuest
 import com.ilsangtech.ilsang.core.model.quest.PopularQuest
 import com.ilsangtech.ilsang.core.model.quest.RecommendedQuest
+import com.ilsangtech.ilsang.core.model.rank.UserRank
 
 sealed interface HomeTabUiState {
     data object Loading : HomeTabUiState
